Fin InitBDD & correction de bug sur les classes

+++ src/main/java/coloc/repository/IChambreRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import coloc.model.*;
 
-public interface IChambreRepository extends JpaRepository<Chambre,Integer>{
+public interface IChambreRepository extends JpaRepository<Chambre,Long>{
 
 	//public List<Chambre> findAllByIdLogement(int idLogement);
 }
