+++ src/main/java/coloc/repository/ILocataireRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import coloc.model.*;
 
-public interface ILocataireRepository extends JpaRepository<Locataire,Integer>{
+public interface ILocataireRepository extends JpaRepository<Locataire,Long>{
 
 	//public Locataire findByIdChambre(int idChambre);
 }
