+++ src/main/java/coloc/repository/ILogementRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import coloc.model.*;
 
-public interface ILogementRepository extends JpaRepository<Logement,Integer>{
+public interface ILogementRepository extends JpaRepository<Logement,Long>{
 
 	//public List<Logement> findAllByIdProprio(int idProprio);
 	//public List<Logement> findAllByAvailabilityWithProprio();
