+++ src/main/java/coloc/repository/IMessageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import coloc.model.*;
 
-public interface IMessageRepository extends JpaRepository<Message,Integer>{
+public interface IMessageRepository extends JpaRepository<Message,Long>{
 
 	//public List<Message> findAllByIdDestinataireWithUtilisateur(int idDestinataire);
 	//public List<Message> findAllByIdEmetteurWithUtilisateur(int idEmetteur);
