+++ src/main/java/coloc/repository/INotationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import coloc.model.*;
 
-public interface INotationRepository extends JpaRepository<Notation,Integer>{
+public interface INotationRepository extends JpaRepository<Notation,Long>{
 
 }
