+++ src/main/java/coloc/repository/IRegleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import coloc.model.*;
 
-public interface IRegleRepository extends JpaRepository<Regle,Integer>{
+public interface IRegleRepository extends JpaRepository<Regle,Long>{
 
 }
