+++ src/main/java/coloc/repository/IUtilisateurRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import coloc.model.*;
 
-public interface IUtilisateurRepository extends JpaRepository<Utilisateur,Integer>{
+public interface IUtilisateurRepository extends JpaRepository<Utilisateur,Long>{
 
 	//public Utilisateur connect(String email, String password);
 	//public Utilisateur findByEmail(String email);
